Change back to c64

+++ pygpcca/_sort_real_schur.py
@@ -95,7 +95,7 @@ def sort_real_schur(
         i for i in range(R.shape[0] + 1) if i not in r + 1
     ]  # construct from them a vector s with the-top left positions of each block
 
-    p = np.empty((len(s) - 1,), dtype=np.complex128)
+    p = np.empty((len(s) - 1,), dtype=np.complex64)
 
     for k in range(1, len(s) - 1):  # debug
         assert R[s[k], s[k] - 1] <= 100 * eps  # debug
